chore: Move the issue deletion
