+++ projects/Mallard/src/helpers/files.ts
@@ -9,18 +9,7 @@ import { getSetting } from './settings'
 import { defaultSettings } from './settings/defaults'
 import { errorService } from 'src/services/errors'
 import { londonTime } from './date'
-import { pushTracking } from 'src/push-notifications/push-tracking'
-import { localIssueListStore } from 'src/hooks/use-issue-on-device'
 import { withCache } from './fetch/cache'
-import { Feature } from 'src/services/logging'
-
-// for cleaning up temporary files when the user hits 'delete all downlods'
-// NOTE: these hard coded names may change when rn-fetch-blob is updated
-const TEMP_FILE_LOCATIONS = [
-    `${RNFetchBlob.fs.dirs.DocumentDir}/`,
-    `${RNFetchBlob.fs.dirs.DocumentDir}/RNFetchBlob_tmp/`,
-]
-const RN_FETCH_TEMP_PREFIX = 'RNFetchBlobTmp'
 
 interface BasicFile {
     filename: string
@@ -50,50 +39,6 @@ export const prepFileSystem = (): Promise<void> =>
     ensureDirExists(FSPaths.issuesDir).then(() =>
         ensureDirExists(`${FSPaths.issuesDir}/daily-edition`),
     )
-
-/**
- * rn-fetch-blob stores the zip file we donwnload in a temporary file. Sometimes,
- * the process that deletes these post extraction fails. This function attempts
- * to clean up such files in a fairly stupid way - just searching for files
- * with RNFetchBlobTmp at the start of the filename
- */
-
-const removeTempFiles = () => {
-    const removeOrphanedTempFiles = async (dir: string) => {
-        try {
-            const isDir = await RNFetchBlob.fs.isDir(dir)
-            if (isDir) {
-                RNFetchBlob.fs.ls(dir).then(files => {
-                    files
-                        .filter(f => f.startsWith(RN_FETCH_TEMP_PREFIX))
-                        .map(f => dir + f)
-                        .map(RNFetchBlob.fs.unlink)
-                })
-            }
-        } catch (error) {
-            await pushTracking(
-                'tempFileRemoveError',
-                JSON.stringify(error),
-                Feature.CLEAR_ISSUES,
-            )
-            console.log(
-                `Error cleaning up temp issue files in directory ${dir}: `,
-                error,
-            )
-            errorService.captureException(error)
-        }
-    }
-    TEMP_FILE_LOCATIONS.forEach(removeOrphanedTempFiles)
-}
-
-export const deleteIssueFiles = async (): Promise<void> => {
-    await RNFetchBlob.fs.unlink(FSPaths.issuesDir)
-    localIssueListStore.reset()
-
-    removeTempFiles()
-
-    await prepFileSystem()
-}
 
 export const getJson = <T extends any>(path: string): Promise<T> =>
     RNFetchBlob.fs.readFile(path, 'utf8').then(d => JSON.parse(d))
@@ -174,14 +119,6 @@ export type DLStatus =
     | { type: 'success' }
     | { type: 'failure' }
 
-const deleteIssue = (localId: string): Promise<void> => {
-    const promise = RNFetchBlob.fs
-        .unlink(FSPaths.issueRoot(localId))
-        .catch(e => errorService.captureException(e))
-    promise.then(() => localIssueListStore.remove(localId))
-    return promise
-}
-
 const withPathPrefix = (prefix: string) => (str: string) => `${prefix}/${str}`
 
 export const getLocalIssues = () =>
@@ -199,21 +136,6 @@ export const issuesToDelete = async (files: string[]) => {
                 .includes(issue) &&
             issue !== `${defaultSettings.contentPrefix}/issues`,
     )
-}
-
-export const clearOldIssues = async (): Promise<void> => {
-    // remove any temp files at this point too
-    removeTempFiles()
-
-    const files = await getLocalIssues()
-
-    const iTD: string[] = await issuesToDelete(files)
-
-    return Promise.all(iTD.map((issue: string) => deleteIssue(issue)))
-        .then(() =>
-            pushTracking('clearOldIssues', 'completed', Feature.CLEAR_ISSUES),
-        )
-        .catch(e => errorService.captureException(e))
 }
 
 export const matchSummmaryToKey = (
