+++ projects/Mallard/src/screens/settings/manage-editions-screen.tsx
@@ -4,7 +4,6 @@ import React from 'react'
 import { Alert, Switch, TouchableOpacity, View, StyleSheet } from 'react-native'
 import { List } from 'src/components/lists/list'
 import { UiBodyCopy } from 'src/components/styled-text'
-import { deleteIssueFiles } from 'src/helpers/files'
 import {
     setMaxAvailableEditions,
     setWifiOnlyDownloads,
@@ -13,6 +12,7 @@ import { WithAppAppearance } from 'src/theme/appearance'
 import { getIssueSummary } from 'src/hooks/use-issue-summary'
 import { sendComponentEvent, ComponentType, Action } from 'src/services/ophan'
 import { MANAGE_EDITIONS_TITLE } from 'src/helpers/words'
+import { deleteIssueFiles } from 'src/download-edition/clear-issues'
 
 const buttonStyles = StyleSheet.create({
     background: {
